requirements.txt
flask-smorest
flask-mongoengine
flask-wtf
flask-login
flask-bcrypt
python-dateutil
email_validator